requirements.txt
colorama==0.4.1
requests==2.16.0
tzlocal==1.5.1
six==1.12.0
progressbar2==3.39.2
setuptools==20.10.1
icalendar==3.11.2
pigpio==1.42
Pillow==5.4.1
spidev==3.2
RPi.GPIO==0.6.5